# perf: speed up MGet/Cache/MSet heplers by grouping commands by connections instead of slots