related to https://github.com/redis/rueidis/issues/278

Single Client Before
```
Benchmark
Benchmark/MGet
Benchmark/MGet-10         	    2770	    520576 ns/op	  310661 B/op	    1005 allocs/op
Benchmark/MGetCache
Benchmark/MGetCache-10    	    6554	    179813 ns/op	  370530 B/op	    1045 allocs/op
PASS
```

Single Client After
```
Benchmark
Benchmark/MGet
Benchmark/MGet-10         	    2749	    490253 ns/op	  310944 B/op	    1006 allocs/op
Benchmark/MGetCache
Benchmark/MGetCache-10    	   11476	    105266 ns/op	  471776 B/op	      55 allocs/op
PASS
```

---

Sentinel Client Before
```
Benchmark
Benchmark/MGet
Benchmark/MGet-10         	     135	   8403304 ns/op	  471670 B/op	    2281 allocs/op
Benchmark/MGetCache
Benchmark/MGetCache-10    	    6852	    150697 ns/op	  369290 B/op	    1042 allocs/op
PASS
```

Sentinel Client After
```
Benchmark
Benchmark/MGet
Benchmark/MGet-10         	    2084	    615179 ns/op	  345916 B/op	    1010 allocs/op
Benchmark/MGetCache
Benchmark/MGetCache-10    	   12031	    100673 ns/op	  469690 B/op	      52 allocs/op
PASS
```

---

Cluster Client Before
```
Benchmark
Benchmark/MGet
Benchmark/MGet-10         	     148	   7580940 ns/op	  454352 B/op	    2113 allocs/op
Benchmark/MGetCache
Benchmark/MGetCache-10    	    6472	    185624 ns/op	  369112 B/op	    1037 allocs/op
PASS
```

Cluster Client After
```
Benchmark
Benchmark/MGet
Benchmark/MGet-10         	    2088	    533159 ns/op	  565037 B/op	    1107 allocs/op
Benchmark/MGetCache
Benchmark/MGetCache-10    	    9366	    130200 ns/op	  563585 B/op	     118 allocs/op
PASS
```

---

Benchmark Source
```go
package sep

import (
	"context"
	"math/rand"
	"testing"
	"time"

	"github.com/redis/rueidis"
)

var charset = []byte("abcdefghijklmnopqrstuvwxyz")

func randstr(n int) string {
	b := make([]byte, n)
	for i := range b {
		b[i] = charset[rand.Intn(len(charset))]
	}
	return string(b)
}

func Benchmark(b *testing.B) {
	makeclient := func() rueidis.Client {
		client, err := rueidis.NewClient(rueidis.ClientOption{
			InitAddress: []string{"127.0.0.1:7001"},
		})
		if err != nil {
			panic(err)
		}
		return client
	}

	// prepare keys
	keys := make([]string, 1000)
	cmds := make(rueidis.Commands, 1000)

	{
		// fill the data
		client := makeclient()
		defer client.Close()

		for i := range keys {
			keys[i] = randstr(10)
			cmds[i] = client.B().Set().Key(keys[i]).Value(randstr(50)).Build()
		}

		for _, resp := range client.DoMulti(context.Background(), cmds...) {
			if err := resp.Error(); err != nil {
				panic(err)
			}
		}
	}

	b.Run("MGet", func(b *testing.B) {
		client := makeclient()
		defer client.Close()
		b.ReportAllocs()
		b.ResetTimer()
		b.RunParallel(func(pb *testing.PB) {
			for pb.Next() {
				ret, err := rueidis.MGet(client, context.Background(), keys)
				if err != nil || len(ret) != len(keys) {
					panic(err)
				}
			}
		})
		b.StopTimer()
	})
	b.Run("MGetCache", func(b *testing.B) {
		client := makeclient()
		defer client.Close()
		b.ReportAllocs()
		b.ResetTimer()
		b.RunParallel(func(pb *testing.PB) {
			for pb.Next() {
				ret, err := rueidis.MGetCache(client, context.Background(), time.Minute, keys)
				if err != nil || len(ret) != len(keys) {
					panic(err)
				}
			}
		})
		b.StopTimer()
	})
}

```